+++ core/src/main/java/arez/SpyImpl.java
@@ -268,15 +268,6 @@ final class SpyImpl
    * {@inheritDoc}
    */
   @Override
-  public boolean isReadOnly( @Nonnull final Observer observer )
-  {
-    return Arez.shouldEnforceTransactionType() && TransactionMode.READ_WRITE != observer.getMode();
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
   public boolean isComputedValue( @Nonnull final Observer observer )
   {
     return observer.isComputedValue();